# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for project and service management with clean architecture",
        "title": "Lucas Web API",
        "termsOfService": "http://localhost:8080/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080/support",
            "email": "support@lucas-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/auth/login-url": {
            "get": {
                "description": "Returns the OAuth2 login URL with optional redirect parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get login URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to redirect after login",
                        "name": "redirect_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}